requirements.txt
Django==2.1.5
Pillow==5.4.1
django-registration-redux==2.2
requests
coverage